+++ fe/lib/components/SeekApiResponse/SeekApiResponse.stories.tsx
@@ -1,13 +1,15 @@
+import type { Meta, StoryObj } from '@storybook/react';
 import { Stack } from 'braid-design-system';
-import React, { type ComponentProps } from 'react';
+import type { ComponentProps } from 'react';
 import { CodeBlock } from 'scoobie';
-import type { Meta, StoryObj } from 'sku/@storybook/react';
 
 import { SeekApiResponse as SeekApiResponseComponent } from './SeekApiResponse';
 
 const Component = (args: ComponentProps<typeof SeekApiResponseComponent>) => (
   <Stack space="large">
-    <CodeBlock language="json">{JSON.stringify(args.data, null, 2)}</CodeBlock>
+    <CodeBlock label="Response" language="json">
+      {JSON.stringify(args.data, null, 2)}
+    </CodeBlock>
     <SeekApiResponseComponent {...args} />
   </Stack>
 );